+++ src/validation.js
@@ -1,26 +1,5 @@
 import { every, filter, isArray, isObject, mapValues, keys } from 'lodash';
 
-export function validate(value, validators) {
-    if (isArray(validators)) {
-        return {
-            isValid: isValid(value, validators),
-            results: getResults(value, validators),
-            errors: getErrors(value, validators)
-        };
-    } else if (isObject(validators)) {
-        return mapValues(
-            validators,
-            (fieldValidators, field) => validate(value && value[field], fieldValidators)
-        );
-    } else {
-        return {
-            isValid: true,
-            results: [],
-            errors: []
-        };
-    }
-}
-
 export function getResults(value, validators, invalidOnly) {
     if (isArray(validators)) {
         const results = validators.map((validator) => validator(value));
@@ -40,21 +19,54 @@ export function getResults(value, validators, invalidOnly) {
     }
 }
 
+export function validate(value, validators) {
+    if (isArray(validators)) {
+        const results = getResults(value, validators);
+
+        return {
+            results,
+            errors: getErrorsFromResults(results),
+            isValid: isValidFromResults(results)
+        };
+    } else if (isObject(validators)) {
+        return mapValues(
+            validators,
+            (fieldValidators, field) => validate(value && value[field], fieldValidators)
+        );
+    } else {
+        return {
+            results: [],
+            errors: [],
+            isValid: true
+        };
+    }
+}
+
 export function getErrors(value, validators) {
     return getResults(value, validators, true);
 }
 
-function everyFieldIsValid(results) {
+export function getErrorsFromResults(results) {
+    if (isArray(results)) {
+        return filter(results, (result) => !result.isValid);
+    } else if (isObject(results)) {
+        return mapValues(results, getErrorsFromResults);
+    } else {
+        return [];
+    }
+}
+
+export function isValidFromResults(results) {
     if (isArray(results)) {
         return every(results, (result) => result.isValid);
     }
 
     return every(keys(results), (field) => {
-        return everyFieldIsValid(results[field]);
+        return isValidFromResults(results[field]);
     });
 }
 
 export function isValid(value, validators) {
     const results = getResults(value, validators);
-    return everyFieldIsValid(results);
+    return isValidFromResults(results);
 }